+++ PrincessGuide/Quest/ViewController/QuestAreaTabViewController.swift
@@ -18,12 +18,14 @@ class QuestAreaTabViewController: TabmanViewController, PageboyViewControllerDat
     override func viewDidLoad() {
         super.viewDidLoad()
         navigationItem.title = NSLocalizedString("Quests", comment: "")
-        viewControllers = [HatsuneEventAreaTableViewController(), ClanBattleTableViewController(), QuestAreaTableViewController(areaType: .normal), QuestAreaTableViewController(areaType: .hard)]
+        viewControllers = [HatsuneEventAreaTableViewController(), ClanBattleTableViewController(), QuestAreaTableViewController(areaType: .normal), QuestAreaTableViewController(areaType: .hard), DungeonBossTableViewController()]
         dataSource = self
         bar.items = [
             Item(title: NSLocalizedString("Event", comment: "")),
             Item(title: NSLocalizedString("Clan Battle", comment: ""))
             ] + [AreaType.normal, .hard].map { Item(title: $0.description) }
+        bar.items?.append(Item(title: NSLocalizedString("Dungeon", comment: "")))
+        
         bar.location = .bottom
 
         ThemeManager.default.apply(theme: Theme.self, to: self) { (themeable, theme) in